+++ protocol-designer/src/load-file/__tests__/fixtures/throughMigrationV0/minimalProtocolOldTransfer.json
